requirements.txt
PyMySQL
Flask
flask-restful
requests
pytz
flask-cors
pyjwt
sqlalchemy
holidays